+++ src/app/Scenes/Consignments/Screens/ConfirmContactInfo.tests.tsx
@@ -1,11 +1,10 @@
 import { waitFor } from "@testing-library/react-native"
-import { PhoneInput } from "app/Components/PhoneInput/PhoneInput"
 import { defaultEnvironment } from "app/relay/createEnvironment"
 import { extractText } from "app/tests/extractText"
 import { flushPromiseQueue } from "app/tests/flushPromiseQueue"
 import { renderWithWrappers, renderWithWrappersTL } from "app/tests/renderWithWrappers"
 import NavigatorIOS from "app/utils/__legacy_do_not_use__navigator-ios-shim"
-import { Button } from "palette"
+import { Button, PhoneInput } from "palette"
 import React from "react"
 import { Alert } from "react-native"
 import { act } from "react-test-renderer"
