+++ src/app/Scenes/Consignments/Screens/ConfirmContactInfo.tsx
@@ -3,14 +3,12 @@ import { captureException } from "@sentry/react-native"
 import { ConfirmContactInfo_me } from "__generated__/ConfirmContactInfo_me.graphql"
 import { ConfirmContactInfoQuery } from "__generated__/ConfirmContactInfoQuery.graphql"
 import { FancyModalHeader } from "app/Components/FancyModal/FancyModalHeader"
-import { PhoneInput } from "app/Components/PhoneInput/PhoneInput"
-import { popToRoot } from "app/navigation/navigate"
 import { defaultEnvironment } from "app/relay/createEnvironment"
 import { updateMyUserProfile } from "app/Scenes/MyAccount/updateMyUserProfile"
 import NavigatorIOS from "app/utils/__legacy_do_not_use__navigator-ios-shim"
 import { renderWithPlaceholder } from "app/utils/renderWithPlaceholder"
 import { useScreenDimensions } from "app/utils/useScreenDimensions"
-import { Box, Input, Spacer, Text } from "palette"
+import { Box, Input, PhoneInput, Spacer, Text } from "palette"
 import React, { useEffect, useRef, useState } from "react"
 import { ActivityIndicator, Alert, BackHandler, ScrollView, View } from "react-native"
 import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
