+++ src/app/Scenes/Consignments/Screens/SubmitArtworkOverview/ContactInformation/ContactInformation.tsx
@@ -1,11 +1,10 @@
 import { captureMessage } from "@sentry/react-native"
 import { ContactInformation_me } from "__generated__/ContactInformation_me.graphql"
 import { ContactInformationQueryRendererQuery } from "__generated__/ContactInformationQueryRendererQuery.graphql"
-import { PhoneInput } from "app/Components/PhoneInput/PhoneInput"
 import { defaultEnvironment } from "app/relay/createEnvironment"
 import { GlobalStore } from "app/store/GlobalStore"
 import { Formik } from "formik"
-import { CTAButton, Flex, Input, Spacer, Text } from "palette"
+import { CTAButton, Flex, Input, PhoneInput, Spacer, Text } from "palette"
 import React, { useState } from "react"
 import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
 import { ErrorView } from "../Components/ErrorView"
