+++ src/app/Scenes/MyAccount/MyAccountEditPhone.tsx
@@ -1,8 +1,8 @@
 import { MyAccountEditPhoneQuery } from "__generated__/MyAccountEditPhoneQuery.graphql"
-import { PhoneInput } from "app/Components/PhoneInput/PhoneInput"
 import { defaultEnvironment } from "app/relay/createEnvironment"
 import { PlaceholderBox } from "app/utils/placeholders"
 import { renderWithPlaceholder } from "app/utils/renderWithPlaceholder"
+import { PhoneInput } from "palette"
 import React, { useEffect, useState } from "react"
 import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
 import { MyAccountEditPhone_me } from "../../../__generated__/MyAccountEditPhone_me.graphql"
