+++ src/app/Scenes/SavedAddresses/SavedAddressesForm.tests.tsx
@@ -1,9 +1,7 @@
 import { SavedAddressesFormTestsQuery } from "__generated__/SavedAddressesFormTestsQuery.graphql"
-import { PhoneInput } from "app/Components/PhoneInput/PhoneInput"
 import { extractText } from "app/tests/extractText"
 import { renderWithWrappers } from "app/tests/renderWithWrappers"
-import { Button, Input } from "palette"
-import { Checkbox } from "palette/elements/Checkbox"
+import { Button, Checkbox, Input, PhoneInput } from "palette"
 import React from "react"
 import { graphql, QueryRenderer } from "react-relay"
 import { createMockEnvironment, MockPayloadGenerator } from "relay-test-utils"
