+++ src/app/Scenes/SavedAddresses/SavedAddressesForm.tsx
@@ -3,9 +3,6 @@ import { SavedAddressesForm_me } from "__generated__/SavedAddressesForm_me.graph
 import { SavedAddressesFormQuery } from "__generated__/SavedAddressesFormQuery.graphql"
 import { CountrySelect } from "app/Components/CountrySelect"
 import { PageWithSimpleHeader } from "app/Components/PageWithSimpleHeader"
-import { cleanUserPhoneNumber } from "app/Components/PhoneInput/cleanUserPhoneNumber"
-import { countryCodes } from "app/Components/PhoneInput/countries"
-import { PhoneInput } from "app/Components/PhoneInput/PhoneInput"
 import { Stack } from "app/Components/Stack"
 import { useToast } from "app/Components/Toast/toastHook"
 import { goBack } from "app/navigation/navigate"
@@ -16,8 +13,10 @@ import { renderWithPlaceholder } from "app/utils/renderWithPlaceholder"
 import { useScreenDimensions } from "app/utils/useScreenDimensions"
 import { Action, action, computed, Computed, createComponentStore } from "easy-peasy"
 import { times } from "lodash"
-import { Flex, Input, Text } from "palette"
-import { Checkbox } from "palette/elements/Checkbox"
+import { Checkbox, Flex, Input, PhoneInput, Text } from "palette"
+import { cleanUserPhoneNumber } from "palette/elements/Input/PhoneInput/cleanUserPhoneNumber"
+import { countryCodes } from "palette/elements/Input/PhoneInput/countries"
+
 import React, { useEffect, useRef, useState } from "react"
 import { Alert } from "react-native"
 import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
