+++ src/palette/elements/Input/PhoneInput/index.ts
@@ -1,0 +1,1 @@
+export * from "./PhoneInput"
