+++ src/palette/elements/Input/index.tsx
@@ -1,2 +1,3 @@
 export * from "./Input"
 export * from "./InputTitle"
+export * from "./PhoneInput"
